+++ tests/models/albert/test_modeling_tf_albert.py
@@ -251,7 +251,7 @@ class TFAlbertModelTest(TFModelTesterMixin, unittest.TestCase):
 
         if return_labels:
             if model_class in get_values(TF_MODEL_FOR_PRETRAINING_MAPPING):
-                inputs_dict["sentence_order_label"] = tf.zeros(self.model_tester.batch_size, dtype=tf.int32)
+                inputs_dict["sentence_order_label"] = tf.zeros(self.model_tester.batch_size, dtype=tf.int64)
 
         return inputs_dict
