+++ tests/models/bart/test_modeling_tf_bart.py
@@ -230,7 +230,7 @@ class TFBartModelTest(TFModelTesterMixin, TFCoreModelTesterMixin, unittest.TestC
 
 
 def _long_tensor(tok_lst):
-    return tf.constant(tok_lst, dtype=tf.int32)
+    return tf.constant(tok_lst, dtype=tf.int64)
 
 
 @require_tf
@@ -238,7 +238,7 @@ class TFBartHeadTests(unittest.TestCase):
     vocab_size = 99
 
     def _get_config_and_data(self):
-        eos_column_vector = tf.ones((4, 1), dtype=tf.int32) * 2
+        eos_column_vector = tf.ones((4, 1), dtype=tf.int64) * 2
         input_ids = tf.concat([ids_tensor((4, 6), self.vocab_size - 3) + 3, eos_column_vector], axis=1)
         batch_size = input_ids.shape[0]
         config = BartConfig(
