+++ tests/models/bert/test_modeling_tf_bert.py
@@ -325,8 +325,8 @@ class TFBertModelTester:
 
         # create attention mask
         half_seq_length = self.seq_length // 2
-        attn_mask_begin = tf.ones((self.batch_size, half_seq_length), dtype=tf.int32)
-        attn_mask_end = tf.zeros((self.batch_size, self.seq_length - half_seq_length), dtype=tf.int32)
+        attn_mask_begin = tf.ones((self.batch_size, half_seq_length), dtype=tf.int64)
+        attn_mask_end = tf.zeros((self.batch_size, self.seq_length - half_seq_length), dtype=tf.int64)
         attn_mask = tf.concat([attn_mask_begin, attn_mask_end], axis=1)
 
         # first forward pass
@@ -349,7 +349,7 @@ class TFBertModelTester:
         # append to next input_ids and
         next_input_ids = tf.concat([input_ids, next_tokens], axis=-1)
         attn_mask = tf.concat(
-            [attn_mask, tf.ones((attn_mask.shape[0], 1), dtype=tf.int32)],
+            [attn_mask, tf.ones((attn_mask.shape[0], 1), dtype=tf.int64)],
             axis=1,
         )
 
@@ -617,7 +617,7 @@ class TFBertModelTest(TFModelTesterMixin, TFCoreModelTesterMixin, unittest.TestC
 
         if return_labels:
             if model_class in get_values(TF_MODEL_FOR_PRETRAINING_MAPPING):
-                inputs_dict["next_sentence_label"] = tf.zeros(self.model_tester.batch_size, dtype=tf.int32)
+                inputs_dict["next_sentence_label"] = tf.zeros(self.model_tester.batch_size, dtype=tf.int64)
 
         return inputs_dict
