+++ tests/models/ctrl/test_modeling_tf_ctrl.py
@@ -233,7 +233,7 @@ class TFCTRLModelLanguageGenerationTest(unittest.TestCase):
     @slow
     def test_lm_generate_ctrl(self):
         model = TFCTRLLMHeadModel.from_pretrained("ctrl")
-        input_ids = tf.convert_to_tensor([[11859, 0, 1611, 8]], dtype=tf.int32)  # Legal the president is
+        input_ids = tf.convert_to_tensor([[11859, 0, 1611, 8]], dtype=tf.int64)  # Legal the president is
         expected_output_ids = [
             11859,
             0,
