+++ tests/models/mbart/test_modeling_tf_mbart.py
@@ -301,7 +301,7 @@ def _assert_tensors_equal(a, b, atol=1e-12, prefix=""):
 
 
 def _long_tensor(tok_lst):
-    return tf.constant(tok_lst, dtype=tf.int32)
+    return tf.constant(tok_lst, dtype=tf.int64)
 
 
 TOLERANCE = 1e-4
