+++ tests/models/openai/test_modeling_tf_openai.py
@@ -257,7 +257,7 @@ class TFOPENAIGPTModelLanguageGenerationTest(unittest.TestCase):
     @slow
     def test_lm_generate_openai_gpt(self):
         model = TFOpenAIGPTLMHeadModel.from_pretrained("openai-gpt")
-        input_ids = tf.convert_to_tensor([[481, 4735, 544]], dtype=tf.int32)  # the president is
+        input_ids = tf.convert_to_tensor([[481, 4735, 544]], dtype=tf.int64)  # the president is
         expected_output_ids = [
             481,
             4735,
