+++ tests/models/roberta/test_modeling_tf_roberta.py
@@ -311,8 +311,8 @@ class TFRobertaModelTester:
 
         # create attention mask
         half_seq_length = self.seq_length // 2
-        attn_mask_begin = tf.ones((self.batch_size, half_seq_length), dtype=tf.int32)
-        attn_mask_end = tf.zeros((self.batch_size, self.seq_length - half_seq_length), dtype=tf.int32)
+        attn_mask_begin = tf.ones((self.batch_size, half_seq_length), dtype=tf.int64)
+        attn_mask_end = tf.zeros((self.batch_size, self.seq_length - half_seq_length), dtype=tf.int64)
         attn_mask = tf.concat([attn_mask_begin, attn_mask_end], axis=1)
 
         # first forward pass
@@ -337,7 +337,7 @@ class TFRobertaModelTester:
         # append to next input_ids and
         next_input_ids = tf.concat([input_ids, next_tokens], axis=-1)
         attn_mask = tf.concat(
-            [attn_mask, tf.ones((attn_mask.shape[0], 1), dtype=tf.int32)],
+            [attn_mask, tf.ones((attn_mask.shape[0], 1), dtype=tf.int64)],
             axis=1,
         )
