+++ tests/models/xlm/test_modeling_tf_xlm.py
@@ -339,7 +339,7 @@ class TFXLMModelLanguageGenerationTest(unittest.TestCase):
     @slow
     def test_lm_generate_xlm_mlm_en_2048(self):
         model = TFXLMWithLMHeadModel.from_pretrained("xlm-mlm-en-2048")
-        input_ids = tf.convert_to_tensor([[14, 447]], dtype=tf.int32)  # the president
+        input_ids = tf.convert_to_tensor([[14, 447]], dtype=tf.int64)  # the president
         expected_output_ids = [
             14,
             447,
